make it run on derecho

+++ dev/scripts/exglobal_atmos_products.sh
@@ -129,7 +129,9 @@ for ((nset = 1; nset <= downset; nset++)); do
     fi
 
     # We are in a loop over downset, save output from mpmd into nset specific output
-    mv mpmd.out "mpmd_${nset}.out"
+    if [[ -f mpmd.out ]]; then
+        mv mpmd.out "mpmd_${nset}.out"
+    fi
 
     # Concatenate grib files from each processor into a single one
     # and clean-up as you go
